+++ identidoc/services/__init__.py
@@ -2,4 +2,4 @@
 
 from .preprocessing import preprocess_file
 from .database import validate_database, ClassificationResultTableRow, insert_record_command, retrieve_records_query
-from .misc import get_current_time_as_POSIX_timestamp, datetime_to_POSIX_timestamp
+from .timehandling import get_current_time_as_POSIX_timestamp, datetime_to_POSIX_timestamp
